% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armodel.R
\name{armodel}
\alias{armodel}
\title{Title armodel}
\usage{
armodel(i)
}
\arguments{
\item{i}{The power you choose for the model.}
}
\value{

}
\description{
This function can use the AR model and see the summary with simplied method.
}
