% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/test1_plot.R
\name{test1_plot}
\alias{test1_plot}
\title{Title}
\usage{
test1_plot(i)

test1_plot(i)
}
\arguments{
\item{i}{this should be the period whose value you want to forecast}
}
\value{



}
\description{
this will plot the prediction value with its prediction interval
from a AR(1) Model
}
