% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2_plot.R
\name{test2_plot}
\alias{test2_plot}
\title{Title test2_plot}
\usage{
test2_plot(i)
}
\arguments{
\item{i}{}
}
\value{

}
\description{
This plots the predicted value with its prediction intervals by
AR(2) model.
}
