% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stationary.R
\name{test_stationary}
\alias{test_stationary}
\title{Title}
\usage{
test_stationary(testdata = data_402)
}
\arguments{
\item{testdata}{}
}
\value{

}
\description{
This function is used to test the staionarity of a series of data by using plots and adf.test.
}
